- changed grading reltaed fields from integer to float - fixing average grading - added fields in activity submissions - submission and review page done for backend - overall score for student in class record done - highest score and class avg score in teacher class page done

+++ database/migrations/2025_02_07_074051_create_items_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
             $table->string('itemName');                  // Title
             $table->text('itemDesc');                    // Description
             $table->enum('itemDifficulty', ['Beginner', 'Intermediate', 'Advanced']);
-            $table->integer('itemPoints');
+            $table->float('itemPoints');
             
             $table->timestamps();
